fix(chatgpt-team): keep the leaderboard's Bronze types, fix its parsing instead
Redeclaring the leaderboard's counters as integer would have created a drift no
deploy could close: its tables exist wherever the connector has run, and
reconcile_bronze_schema.py only ADDs columns — a differing type is reported and
left alone, because MODIFY COLUMN rewrites data and is an operator's decision.
A fresh install would get Int64 and every existing one would stay on Decimal.
So n_threads, n_turns, current_streak, text_tokens and lines_added go back to
`number`, in the manifest and in the snapshot alike. The new sessions stream
keeps Int64: it has no table anywhere yet.

The reading is corrected at the dbt boundary, where it belongs. Measured on the
server this runs against: toString of a Decimal trims its trailing zeros, so
toUInt32OrNull(toString(...)) returns a whole reading unchanged and yields NULL
for a fractional or negative one — the conversion was already exact. What was
wrong sat around it. coalesce(..., 0) turned that NULL into a zero, reporting a
measurement of none where there was no measurement, so the JSON blob now
carries '' for a value that did not parse, as it already did for the sessions
counters beside it.

session_count and lines_added keep their zero, and the comment now says why:
the class contract forbids NULL there. Neither can admit a row on it — the
emission filter tests the same expression — and conversation_count is nullable
and keeps the distinction intact.

total_users needed nothing: every value it holds parses, for the same
trailing-zero reason.

+++ src/ingestion/scripts/connectors-ddl/chatgpt-team.sql
@@ -120,11 +120,11 @@ CREATE TABLE IF NOT EXISTS bronze_chatgpt_team.chatgpt_team_codex_user_daily
     `user_id` Nullable(String),
     `name` Nullable(String),
     `credits` Nullable(Decimal(38, 9)),
-    `n_threads` Nullable(Int64),
-    `n_turns` Nullable(Int64),
-    `current_streak` Nullable(Int64),
-    `text_tokens` Nullable(Int64),
-    `lines_added` Nullable(Int64)
+    `n_threads` Nullable(Decimal(38, 9)),
+    `n_turns` Nullable(Decimal(38, 9)),
+    `current_streak` Nullable(Decimal(38, 9)),
+    `text_tokens` Nullable(Decimal(38, 9)),
+    `lines_added` Nullable(Decimal(38, 9))
 )
 ENGINE = ReplacingMergeTree(_airbyte_extracted_at)
 ORDER BY unique_key
